receiver/requirements.txt
connexion[flask,uvicorn,swagger-ui]>=2.14.2
pykafka
flask-cors
PyYAML
setuptools
